Ordeno archivo singIn llevo funcion registrar
