--- src/lib/getHash.js
@@ -1,4 +1,0 @@
-const getHash = (hash) => {
-  console.log(hash.slice(1).toLocaleLowerCase().split('/')[1]);
-  // window location
-};
