+++ src/main.js
@@ -1,7 +1,8 @@
 /* eslint-disable spaced-comment */
 // Este es el punto de entrada de tu aplicacion
 import { changePages } from './routes/router.js';
-import { logueo, register } from './pages/login.js';
+import { logueo } from './pages/login.js';
+import { register } from './pages/signIn.js';
 
 const init = () => {
   changePages(window.location.hash);
